+++ wgsl_to_wgpu/src/data/globals/vertex_fragment.rs
@@ -221,9 +221,114 @@ pub mod depth_texture_msaa {
         }
     }
 }
+pub mod color_texture_array_2d {
+    pub const GROUP: u32 = 0u32;
+    pub const BINDING: u32 = 11u32;
+    pub const LAYOUT: wgpu::BindGroupLayoutEntry = wgpu::BindGroupLayoutEntry {
+        binding: BINDING,
+        visibility: wgpu::ShaderStages::all(),
+        ty: wgpu::BindingType::Texture {
+            sample_type: wgpu::TextureSampleType::Float { filterable: true },
+            view_dimension: wgpu::TextureViewDimension::D2,
+            multisampled: false,
+        },
+        count: None,
+    };
+    pub type Resource<'a> = &'a wgpu::TextureView;
+    pub fn bind_group_entry(resource: Resource) -> wgpu::BindGroupEntry<'_> {
+        wgpu::BindGroupEntry {
+            binding: BINDING,
+            resource: wgpu::BindingResource::TextureView(resource),
+        }
+    }
+}
+pub mod color_texture_array_cube {
+    pub const GROUP: u32 = 0u32;
+    pub const BINDING: u32 = 12u32;
+    pub const LAYOUT: wgpu::BindGroupLayoutEntry = wgpu::BindGroupLayoutEntry {
+        binding: BINDING,
+        visibility: wgpu::ShaderStages::all(),
+        ty: wgpu::BindingType::Texture {
+            sample_type: wgpu::TextureSampleType::Float { filterable: true },
+            view_dimension: wgpu::TextureViewDimension::Cube,
+            multisampled: false,
+        },
+        count: None,
+    };
+    pub type Resource<'a> = &'a wgpu::TextureView;
+    pub fn bind_group_entry(resource: Resource) -> wgpu::BindGroupEntry<'_> {
+        wgpu::BindGroupEntry {
+            binding: BINDING,
+            resource: wgpu::BindingResource::TextureView(resource),
+        }
+    }
+}
 pub mod transforms {
     pub const GROUP: u32 = 1u32;
     pub const BINDING: u32 = 0u32;
+    pub const LAYOUT: wgpu::BindGroupLayoutEntry = wgpu::BindGroupLayoutEntry {
+        binding: BINDING,
+        visibility: wgpu::ShaderStages::all(),
+        ty: wgpu::BindingType::Buffer {
+            ty: wgpu::BufferBindingType::Uniform,
+            has_dynamic_offset: false,
+            min_binding_size: None,
+        },
+        count: None,
+    };
+    pub type Resource<'a> = wgpu::BufferBinding<'a>;
+    pub fn bind_group_entry(resource: Resource) -> wgpu::BindGroupEntry<'_> {
+        wgpu::BindGroupEntry {
+            binding: BINDING,
+            resource: wgpu::BindingResource::Buffer(resource),
+        }
+    }
+}
+pub mod scalar {
+    pub const GROUP: u32 = 1u32;
+    pub const BINDING: u32 = 1u32;
+    pub const LAYOUT: wgpu::BindGroupLayoutEntry = wgpu::BindGroupLayoutEntry {
+        binding: BINDING,
+        visibility: wgpu::ShaderStages::all(),
+        ty: wgpu::BindingType::Buffer {
+            ty: wgpu::BufferBindingType::Uniform,
+            has_dynamic_offset: false,
+            min_binding_size: None,
+        },
+        count: None,
+    };
+    pub type Resource<'a> = wgpu::BufferBinding<'a>;
+    pub fn bind_group_entry(resource: Resource) -> wgpu::BindGroupEntry<'_> {
+        wgpu::BindGroupEntry {
+            binding: BINDING,
+            resource: wgpu::BindingResource::Buffer(resource),
+        }
+    }
+}
+pub mod vector {
+    pub const GROUP: u32 = 1u32;
+    pub const BINDING: u32 = 2u32;
+    pub const LAYOUT: wgpu::BindGroupLayoutEntry = wgpu::BindGroupLayoutEntry {
+        binding: BINDING,
+        visibility: wgpu::ShaderStages::all(),
+        ty: wgpu::BindingType::Buffer {
+            ty: wgpu::BufferBindingType::Uniform,
+            has_dynamic_offset: false,
+            min_binding_size: None,
+        },
+        count: None,
+    };
+    pub type Resource<'a> = wgpu::BufferBinding<'a>;
+    pub fn bind_group_entry(resource: Resource) -> wgpu::BindGroupEntry<'_> {
+        wgpu::BindGroupEntry {
+            binding: BINDING,
+            resource: wgpu::BindingResource::Buffer(resource),
+        }
+    }
+}
+pub mod matrix {
+    pub const GROUP: u32 = 1u32;
+    pub const BINDING: u32 = 3u32;
     pub const LAYOUT: wgpu::BindGroupLayoutEntry = wgpu::BindGroupLayoutEntry {
         binding: BINDING,
         visibility: wgpu::ShaderStages::all(),