+++ CHANGELOG.md
@@ -6,12 +6,18 @@ Breaking changes in the generated code will be considered as breaking changes wh
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## unreleased
+## 0.10.1 - 2024-09-02
+### Added
+* Added support for scalar, vector, and matrix bindings.
+* Added support for texture arrays.
+
+## 0.10.0 - 2024-08-29
 ### Added
 * Added a `rustfmt` option for formatting the output code with the executable used for `cargo fmt`.
 
 ### Changed
 * Changed generated bind groups and bind group layouts to use labels with the group index like "BindGroup0" or "BindGroupLayout0".
+* Changed bind group set methods to be generic over the pass type to better support mixed entry function stages.
 
 ### Fixed
 * Fixed an issue where vertex entry code would only be generated for the first vertex entry.